inference/requirements.txt
absl-py==0.12.0
apache-beam==2.34.0
appdirs==1.4.4
argon2-cffi==21.1.0
astroid==2.8.5
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
avro-python3==1.9.2.1
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
clang==5.0
colorama==0.4.4
contextlib2==21.6.0
crcmod==1.7
cycler==0.11.0
Cython==0.29.24
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.6
docopt==0.6.2
docutils==0.18
entrypoints==0.3
fastavro==1.4.7
fire==0.4.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.5.0
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-bigquery==2.30.1
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
h5py==3.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.8.2
imutils==0.5.4
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==23.2.1
kiwisolver==1.3.2
labelImg==1.8.6
lazy-object-proxy==1.6.0
libclang==12.0.0
librosa==0.8.1
llvmlite==0.37.0
lvis==0.5.3
lxml==4.6.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
neural-structured-learning==1.3.1
notebook==6.4.5
numba==0.54.1
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
object-detection @ file:///C:/Users/Nathan/Documents/Dev/tensorflow-models/research
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
opt-einsum==3.3.0
orjson==3.6.4
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.7.1
platformdirs==2.4.0
pooch==1.5.2
portalocker==2.3.2
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.22
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
py-cpuinfo==8.0.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.8.1
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pycparser==2.21
pydot==1.4.2
Pygments==2.10.0
pylint==2.11.1
pymongo==3.12.1
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pywin32==302
pywin32-ctypes==0.2.0
pywinpty==1.1.5
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
readme-renderer==30.0
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
rfc3986==1.5.0
rsa==4.7.2
sacrebleu==2.0.0
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
simple-pid==1.0.1
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
tabulate==0.8.9
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.15.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.9.0
tensorflow-io==0.22.0
tensorflow-io-gcs-filesystem==0.22.0
tensorflow-metadata==1.4.0
tensorflow-model-optimization==0.7.0
tensorflow-text==2.7.0
tensorflowjs==3.11.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-support==0.3.0
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
twine==3.6.0
typeguard==2.13.0
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
wrapt==1.12.1
zipp==3.6.0
